+++ packages/atproto/lib/src/atproto.dart
@@ -6,9 +6,9 @@
 import 'package:atproto_core/atproto_core.dart' as core;
 
 import 'atproto_service.dart';
+import 'entities/session.dart';
 import 'handles/handles_service.dart';
 import 'repositories/repositories_service.dart';
-import 'sessions/session.dart';
 import 'sessions/sessions_service.dart';
 
 abstract class ATProto {
