+++ packages/atproto/lib/src/entities/current_session.freezed.dart
