+++ packages/atproto/lib/src/entities/session.freezed.dart
